+++ docs/Whats-New.md
@@ -154,6 +154,7 @@ Vanilla fixes:
 - Fix a bug where name (hover-on tooltip) of friendly spies was displayed as the disguise's name (by Rampastring)
 - Fix an exploit where hijacked build-limit units could be deployed to erase the hijacker and circumvent the build limits of both the hijacker and its target unit type (by Rampastring)
 - Fix an edge case crash when AI is attempting to find a location to place a structure to (by Rampastring)
+- Fix a bug that allowed players to build objects they are not normally allowed to build through crafted network requests (by Rampastring)
 
 :::